firmware: Add 'reset' command
Introduce 'reset' command that re-initializes the device.

Note: as the ATmega32u4 does USB-to-serial on its own, resetting the chip also
re-initializes USB and therefore makes the USB TTY disappear for a short period
of time.

This has the disadvantage of disconnecting serial clients but the benefit of
knocking a blocked USB stack loose along the way.